refactor out Pool module
