+++ server/ghc-persistent-worker-server.cabal
@@ -11,13 +11,11 @@ maintainer:         ianwookim@gmail.com
 category:           Development
 build-type:         Simple
 extra-doc-files:    CHANGELOG.md
--- extra-source-files:
-
-common warnings
-    ghc-options: -Wall
 
 executable Server
     main-is:          Server.hs
+    ghc-options:      -Wall -Werror
+    other-modules:    Pool
     hs-source-dirs:   app
     build-depends:    base >=4.19,
                       binary,
